GO-2512 Update backlinks on links creation
